refactor: 역 생성 서비스 분리
