+++ src/main/java/subway/exception/StationsAlreadyExistException.java
@@ -5,7 +5,7 @@ import subway.error.exception.SubwayException;
 
 public class StationsAlreadyExistException extends SubwayException {
 
-    public static SubwayException EXECUTE = new StationsAlreadyExistException();
+    public static SubwayException THROW = new StationsAlreadyExistException();
 
     public StationsAlreadyExistException() {
         super(new ErrorCode(400, "입력한 역들이 모두 존재합니다."));
